palm_model_system-v23.10-rc.1/packages/static_driver/palm_csd/requirements.txt
netCDF4==1.6.2
rasterio==1.3.7
scipy==1.9.3
pyproj==3.5.0
pyyaml==6.0
